+++ packages/catalog-client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/catalog-client
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies [08835a61d]
+- Updated dependencies [a9fd599f7]
+- Updated dependencies [bcc211a08]
+  - @backstage/catalog-model@0.4.0
+
 ## 0.3.1
 
 ### Patch Changes